+++ goldens/public-api/platform-browser-dynamic/index.api.md
@@ -4,10 +4,11 @@
 
 ```ts
 
-import * as _angular_core from '@angular/core';
 import { Compiler } from '@angular/core';
 import { CompilerFactory } from '@angular/core';
 import { CompilerOptions } from '@angular/core';
+import * as i0 from '@angular/core';
+import { StaticProvider } from '@angular/core';
 import { Version } from '@angular/core';
 
 // @public @deprecated (undocumented)
@@ -17,7 +18,7 @@ export class JitCompilerFactory implements CompilerFactory {
 }
 
 // @public (undocumented)
-export const platformBrowserDynamic: (extraProviders?: _angular_core.StaticProvider[]) => _angular_core.PlatformRef;
+export const platformBrowserDynamic: (extraProviders?: StaticProvider[]) => i0.PlatformRef;
 
 // @public (undocumented)
 export const VERSION: Version;
