+++ goldens/public-api/platform-server/testing/index.api.md
@@ -6,9 +6,10 @@
 
 import * as i0 from '@angular/core';
 import * as i1 from '@angular/platform-browser-dynamic/testing';
+import { StaticProvider } from '@angular/core';
 
 // @public
-export const platformServerTesting: (extraProviders?: i0.StaticProvider[]) => i0.PlatformRef;
+export const platformServerTesting: (extraProviders?: StaticProvider[]) => i0.PlatformRef;
 
 // @public
 export class ServerTestingModule {
