+++ packages/platform-browser-dynamic/src/platform-browser-dynamic.ts
@@ -6,20 +6,8 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {createPlatformFactory} from '@angular/core';
-
-import {platformCoreDynamic} from './platform_core_dynamic';
-import {INTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS} from './platform_providers';
-
-export * from './private_export';
 export {VERSION} from './version';
 export {JitCompilerFactory} from './compiler_factory';
+export {platformBrowserDynamic} from './platform_providers';
 
-/**
- * @publicApi
- */
-export const platformBrowserDynamic = createPlatformFactory(
-  platformCoreDynamic,
-  'browserDynamic',
-  INTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS,
-);
+export * from './private_export';
