--- packages/platform-browser-dynamic/src/platform_core_dynamic.ts
@@ -1,26 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.dev/license
- */
-
-import {
-  COMPILER_OPTIONS,
-  CompilerFactory,
-  createPlatformFactory,
-  platformCore,
-} from '@angular/core';
-
-import {JitCompilerFactory} from './compiler_factory';
-
-/**
- * A platform that included corePlatform and the compiler.
- *
- * @publicApi
- */
-export const platformCoreDynamic = createPlatformFactory(platformCore, 'coreDynamic', [
-  {provide: COMPILER_OPTIONS, useValue: {}, multi: true},
-  {provide: CompilerFactory, useClass: JitCompilerFactory, deps: [COMPILER_OPTIONS]},
-]);
