--- packages/platform-browser-dynamic/testing/src/platform_core_dynamic_testing.ts
@@ -1,18 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.dev/license
- */
-
-import {createPlatformFactory, PlatformRef} from '@angular/core';
-import {ɵplatformCoreDynamic as platformCoreDynamic} from '@angular/platform-browser-dynamic';
-
-/**
- * Platform for dynamic tests
- *
- * @publicApi
- */
-export const platformCoreDynamicTesting: (extraProviders?: any[]) => PlatformRef =
-  createPlatformFactory(platformCoreDynamic, 'coreDynamicTesting', []);
