--- packages/platform-browser-dynamic/testing/src/private_export_testing.ts
@@ -1,10 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.dev/license
- */
-
-export {platformCoreDynamicTesting as ɵplatformCoreDynamicTesting} from './platform_core_dynamic_testing';
-export {ɵDOMTestComponentRenderer} from '@angular/platform-browser/testing';
