+++ packages/platform-browser-dynamic/testing/src/testing.ts
@@ -7,20 +7,15 @@
  */
 
 import {createPlatformFactory, NgModule, PlatformRef, StaticProvider} from '@angular/core';
-import {ɵINTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS as INTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS} from '@angular/platform-browser-dynamic';
+import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 import {BrowserTestingModule} from '@angular/platform-browser/testing';
-
-import {platformCoreDynamicTesting} from './platform_core_dynamic_testing';
-
-export * from './private_export_testing';
 
 /**
  * @publicApi
  */
 export const platformBrowserDynamicTesting = createPlatformFactory(
-  platformCoreDynamicTesting,
+  platformBrowserDynamic,
   'browserDynamicTesting',
-  INTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS,
 );
 
 /**
