+++ packages/platform-server/testing/BUILD.bazel
@@ -9,6 +9,7 @@ ng_module(
     srcs = glob(["**/*.ts"]),
     deps = [
         "//packages/core",
+        "//packages/platform-browser-dynamic",
         "//packages/platform-browser-dynamic/testing",
         "//packages/platform-server",
     ],
