+++ packages/platform-server/testing/src/server.ts
@@ -6,15 +6,18 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {createPlatformFactory, NgModule} from '@angular/core';
-import {
-  BrowserDynamicTestingModule,
-  ɵplatformCoreDynamicTesting as platformCoreDynamicTesting,
-} from '@angular/platform-browser-dynamic/testing';
+import {createPlatformFactory, NgModule, platformCore, StaticProvider} from '@angular/core';
+import {ɵINTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS as INTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS} from '@angular/platform-browser-dynamic';
+import {BrowserDynamicTestingModule} from '@angular/platform-browser-dynamic/testing';
 import {
   ɵINTERNAL_SERVER_PLATFORM_PROVIDERS as INTERNAL_SERVER_PLATFORM_PROVIDERS,
   ɵSERVER_RENDER_PROVIDERS as SERVER_RENDER_PROVIDERS,
 } from '@angular/platform-server';
+
+const INTERNAL_SERVER_DYNAMIC_PLATFORM_TESTING_PROVIDERS: StaticProvider[] = [
+  ...INTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS,
+  ...INTERNAL_SERVER_PLATFORM_PROVIDERS,
+];
 
 /**
  * Platform for testing
@@ -22,9 +25,9 @@ import {
  * @publicApi
  */
 export const platformServerTesting = createPlatformFactory(
-  platformCoreDynamicTesting,
+  platformCore,
   'serverTesting',
-  INTERNAL_SERVER_PLATFORM_PROVIDERS,
+  INTERNAL_SERVER_DYNAMIC_PLATFORM_TESTING_PROVIDERS,
 );
 
 /**
